fix: renamed file to reflect new functionality
