+++ packages/react/src/themes/default-theme.ts
@@ -1,0 +1,64 @@
+import { Theme } from './default-types';
+import { generateTokens } from './tokens-generator';
+
+export const colors: Theme['colors'] = {
+    'primary-1.1': '#006296',
+    'primary-1.2': '#84C6EA',
+    'primary-1.3': '#003A5A',
+    'primary-1.4': '#E0F0F9',
+    'primary-2': '#012639',
+    'primary-3': '#004E78',
+    'secondary-4.1': '#EF483E',
+    'secondary-4.2': '#F9B6B2',
+    'secondary-4.3': '#D41F14',
+    'brand-05': '#E0F0F9',
+    'brand-20': '#84C6EA',
+    'brand-50': '#006296',
+    'brand-70': '#003A5A',
+    'brand-80': '#012639',
+    'accent-20': '#F9B6B2',
+    'accent-50': '#EF483E',
+    'accent-70': '#D41F14',
+    white: '#FFFFFF',
+    'neutral-02': '#FAFAFA',
+    'neutral-05': '#F1F2F2',
+    'neutral-15': '#DBDEE1',
+    'neutral-30': '#B7BBC2',
+    'neutral-50': '#878F9A',
+    'neutral-65': '#60666E',
+    'neutral-90': '#1B1C1E',
+    'colored-white': '#FAFAFA',
+    'light-grey': '#F1F2F2',
+    grey: '#DBDEE1',
+    'mid-grey': '#B7BBC2',
+    'dark-grey': '#60666E',
+    black: '#000000',
+    'info-1.1': '#602FA0',
+    'success-1.1': '#008533',
+    'success-1.2': '#F6FCF8',
+    'success-1.3': '#8ADDA9',
+    'alert-2.1': '#CD2C23',
+    'alert-2.2': '#FFFAFB',
+    'warning-3.1': '#F5A200',
+    'warning-3.2': '#FFF9F5',
+    'warning-3.3': '#FFB302',
+    'warning-3.4': '#A36D00',
+    'success-05': '#F6FCF8',
+    'success-20': '#8ADDA9',
+    'success-50': '#008533',
+    'success-70': '#004F1E',
+    'alert-05': '#FDF7F6',
+    'alert-20': '#F99D99',
+    'alert-50': '#CD2C23',
+    'alert-70': '#7B1A15',
+    'warning-05': '#FFFBF5',
+    'warning-20': '#FFDD99',
+    'warning-50': '#F5A200',
+    'warning-70': '#9E6900',
+    'informative-05': '#F9F7FB',
+    'informative-20': '#CFC1E3',
+    'informative-50': '#602FA0',
+    'informative-70': '#3A1C60',
+};
+
+export const tokens: Theme['tokens'] = generateTokens(colors);
